specify 'main' as base branch for Renovate
The default branch for this repository is 'release' to accommodate `pre-commit
autoupdate`.  However, the intended merge base for development and updates is
'main'.

This commit configures Renovate to explicitly use 'main' as the base branch for
its operations, ensuring that pull requests are correctly targeted and based on
the 'main' branch rather than 'release'.